data-ingestion/requirements.txt
pandas>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
pinecone>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pytest>=7.0.0
python-dateutil>=2.8.2 